// src/api/models/get-signable-transfer-request-v1.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Immutable X API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 * Contact: support@immutable.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { SignableToken } from './signable-token';

/**
 * 
 * @export
 * @interface GetSignableTransferRequestV1
 */
export interface GetSignableTransferRequestV1 {
    /**
     * Amount of the token to transfer
     * @type {string}
     * @memberof GetSignableTransferRequestV1
     */
    'amount': string;
    /**
     * Ethereum address of the receiving user
     * @type {string}
     * @memberof GetSignableTransferRequestV1
     */
    'receiver': string;
    /**
     * Ethereum address of the transferring user
     * @type {string}
     * @memberof GetSignableTransferRequestV1
     */
    'sender': string;
    /**
     * 
     * @type {SignableToken}
     * @memberof GetSignableTransferRequestV1
     */
    'token': SignableToken;
}